code needs some correction :anger:

+++ posSystem/src/main/java/tgr/databaseControl/dbRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface dbRepository extends CrudRepository<dbControl, Long> {
+public interface dbRepository extends CrudRepository<dbControl, Integer> {
     //dbControl findData(String OrderID, String LastName);
 }
